Rejigger stuff so that on-disk ACLs actually work + support Markdown posts
This took a long time.
I stressed out over it because the code to load perms from disk duplicates functionality that's in ACL already.
/ideally/ I would could reuse it like
```
for rule in open("cooking-lessons.acl"):
	if rule[0] == "+":
		ACL.allow(rule[1:], ("post_view", path="cooking-lessons"))
	if rule[0] == "-":
		ACL.deny(rule[1:], ("post_view", path="cooking-lessons"))
```
and then ACL.check()'s logic could make sure the rules are applied correctly.

But I can't do this. The way I have ACL setup rules are attached to endpoints,
and they have to be attached to endpoints in order for routing to find them, don't they?
but these rules need to be parametric.
Bah
I don't know
I wish someone would take a second look at this and give me wisdom